requirements.txt
requests
numpy
Django==2.2.7
djangorestframework